Add Sidemenu to Exchange page. Fix Profile not shown.
