+++ src/common/menu.js
@@ -1,5 +1,6 @@
-import { isUrl } from '../utils/utils';
+import { formatter } from './menufuncs';
 import { menuData as projectMenuData } from './projectmenu';
+import { menuData as dexMenuData } from './exchangemenu';
 
 const menuData = [
     {
@@ -32,14 +33,10 @@ const menuData = [
         path: 'portfolio',
     },
     {
-        name: 'exchange',
-        icon: 'exchange',
-        path: 'exchange',
-    },
-    {
-        name: 'wallet',
-        icon: 'wallet',
-        path: 'wallet',
+        name: 'dex',
+        icon: 'dex',
+        path: 'dex',
+        children: dexMenuData,
     },
     {
         name: 'account',
@@ -60,28 +57,6 @@ const menuData = [
         ],
     },
 ];
-
-function formatter(data, parentPath = '/', parentAuthority) {
-    return data.map(item => {
-        let { path } = item;
-        if (!isUrl(path)) {
-            path = parentPath + item.path;
-        }
-        const result = {
-            ...item,
-            path,
-            authority: item.authority || parentAuthority,
-        };
-        if (item.children) {
-            result.children = formatter(
-                item.children,
-                `${parentPath}${item.path}/`,
-                item.authority
-            );
-        }
-        return result;
-    });
-}
 
 export const getMenuData = () => formatter(menuData);
