+++ src/common/menufuncs.js
@@ -1,0 +1,23 @@
+import { isUrl } from '../utils/utils';
+
+export function formatter(data, parentPath = '/', parentAuthority) {
+    return data.map(item => {
+        let { path } = item;
+        if (!isUrl(path)) {
+            path = parentPath + item.path;
+        }
+        const result = {
+            ...item,
+            path,
+            authority: item.authority || parentAuthority,
+        };
+        if (item.children) {
+            result.children = formatter(
+                item.children,
+                `${parentPath}${item.path}/`,
+                item.authority
+            );
+        }
+        return result;
+    });
+}
