+++ src/scenes/Exchange/components/ExchangeSider/components/ExchangeMenu/ExchangeMenu.less
@@ -1,0 +1,84 @@
+@import '~antd/lib/style/themes/default.less';
+@ease-in-out-circ: cubic-bezier(0.78, 0.14, 0.15, 0.86);
+.logo {
+    height: 64px;
+    position: relative;
+    line-height: 64px;
+    padding-left: (@menu-collapsed-width - 32px) / 2;
+    transition: all 0.3s;
+    background: #002140;
+    overflow: hidden;
+    img {
+        display: inline-block;
+        vertical-align: middle;
+        height: 32px;
+    }
+    h1 {
+        color: white;
+        display: inline-block;
+        vertical-align: middle;
+        font-size: 20px;
+        margin: 0 0 0 12px;
+        font-family: 'Myriad Pro', 'Helvetica Neue', Arial, Helvetica, sans-serif;
+        font-weight: 600;
+    }
+}
+
+.sider {
+    min-height: 100vh;
+    box-shadow: 2px 0 6px rgba(0, 21, 41, 0.35);
+    position: relative;
+    z-index: 10;
+    background-image: linear-gradient(180deg, #225f63, #19473a);
+
+    .menu_title {
+        margin: 16px 0;
+        text-align: center;
+        vertical-align: baseline;
+        color: #dedede;
+        font-weight: 500;
+    }
+}
+
+.icon {
+    width: 14px;
+    margin-right: 10px;
+}
+
+.banner {
+    width: 100%;
+}
+
+:global {
+    .drawer .drawer-content {
+        background: #001529;
+    }
+
+    .exchange-side-menu {
+        .ant-menu {
+            .ant-menu-submenu,
+            .ant-menu-item > a {
+                color: #dedede;
+            }
+        }
+        .ant-menu-inline-collapsed {
+            & > .ant-menu-item .sider-menu-item-img + span,
+            &
+                > .ant-menu-item-group
+                > .ant-menu-item-group-list
+                > .ant-menu-item
+                .sider-menu-item-img
+                + span,
+            & > .ant-menu-submenu > .ant-menu-submenu-title .sider-menu-item-img + span {
+                max-width: 0;
+                display: inline-block;
+                opacity: 47%;
+            }
+        }
+        .ant-menu-item .sider-menu-item-img + span,
+        .ant-menu-submenu-title .sider-menu-item-img + span {
+            transition: opacity 0.3s @ease-in-out, width 0.3s @ease-in-out;
+            opacity: 1;
+        }
+    }
+}
