+++ src/scenes/Exchange/components/XchangeTransfer/XchangeTransfer.less
@@ -1,0 +1,78 @@
+.header {
+    background-color: #dcf0f1;
+}
+
+.container {
+    margin: 24px;
+    height: 100%;
+
+    .overview {
+        height: 100%;
+    }
+
+    .padbottom {
+        padding-bottom: 24px;
+    }
+}
+.balancetitle {
+    padding-bottom: 10px;
+}
+.balance {
+    text-align: center;
+    min-height: 100px;
+    vertical-align: middle;
+    padding: 20px 10px;
+    font-weight: 600;
+    margin: 5px;
+    border: solid 1px #dcf0f1;
+    -webkit-box-shadow: 5px 5px 15px -5px rgba(166, 166, 166, 1);
+    -moz-box-shadow: 5px 5px 15px -5px rgba(166, 166, 166, 1);
+    box-shadow: 5px 5px 15px -5px rgba(166, 166, 166, 1);
+}
+.balancespan {
+    background-color: #dcf0f1;
+}
+.transfertitle {
+    padding-top: 30px;
+}
+.currenciesicon {
+    width: 30px;
+    height: 30px;
+}
+.transferamount {
+    width: 100%;
+    min-width: 200px;
+    margin: 0 8px;
+}
+.btnexchange {
+    padding-left: 50px;
+    padding-right: 50px;
+    height: 50px;
+    background-color: green;
+}
+@media (max-width: 1600px) {
+    .container {
+        margin: 24px;
+    }
+}
+
+@media (max-width: 1200px) {
+    .container {
+        margin: 24px;
+    }
+}
+
+@media (max-width: 992px) {
+    .container {
+        margin: 24px;
+    }
+}
+
+@media (max-width: 576px) {
+    .container {
+        margin: 12px;
+        .overviewchart {
+            margin-bottom: 12px;
+        }
+    }
+}
